requirements.txt
torch
transformers
sentence-transformers
scikit-learn
numpy
pymupdf
pandas
opencv-python
pytesseract
torchvision
Pillow
